move assign and assignGrad ops to tf package
