+++ spark/dl/src/main/scala/com/intel/analytics/bigdl/nn/tf/StateOps.scala
@@ -13,14 +13,41 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.intel.analytics.bigdl.nn.ops
+package com.intel.analytics.bigdl.nn.tf
 
 import com.intel.analytics.bigdl.nn.abstractnn.Activity
+import com.intel.analytics.bigdl.nn.ops.Operation
+import com.intel.analytics.bigdl.tensor.Tensor
 import com.intel.analytics.bigdl.tensor.TensorNumericMath.{NumericWildcard, TensorNumeric}
-import com.intel.analytics.bigdl.tensor._
-import com.intel.analytics.bigdl.utils.Table
+import com.intel.analytics.bigdl.utils.{T, Table}
 
 import scala.reflect.ClassTag
+
+
+private[bigdl] class Variable[T: ClassTag](
+  val variableValue: Tensor[T],
+  val variableGradient: Tensor[T]
+)(implicit ev: TensorNumeric[T])
+  extends Operation[Activity, Tensor[T], T] with WithoutInput{
+
+  override def clearState(): this.type = {
+    this
+  }
+
+  override def parameters(): (Array[Tensor[T]], Array[Tensor[T]]) = {
+    (Array(this.variableValue), Array(this.variableGradient))
+  }
+
+  override def updateOutput(input: Activity): Tensor[T] = {
+    this.output.resizeAs(variableValue)
+    this.output.copy(variableValue)
+    output
+  }
+
+  override def accGradParameters(input: Activity, gradOutput: Tensor[T]): Unit = {
+    this.variableGradient.add(ev.fromType[Double](1.0), gradOutput)
+  }
+}
 
 /**
  * Update 'ref' by assigning 'value' to it.
@@ -41,7 +68,7 @@ import scala.reflect.ClassTag
  *
  * @tparam T Numeric type. Only support float/double now
  */
-class Assign[T: ClassTag](
+private[bigdl] class Assign[T: ClassTag](
   val validateShape: Boolean = true,
   val useLocking: Boolean = true
 )
@@ -76,7 +103,11 @@ class Assign[T: ClassTag](
   }
 }
 
-object Assign {
-  def apply[T: ClassTag]()(implicit ev: TensorNumeric[T]): Operation[Activity, Activity, T]
-  = ModuleToOperation[T](new Assign())
+private[bigdl] class AssignGrad[T: ClassTag](grad: Tensor[T])(implicit ev: TensorNumeric[T])
+  extends Operation[Tensor[T], Activity, T]{
+
+  override def updateOutput(input: Tensor[T]): Activity = {
+    grad.copy(input)
+    null
+  }
 }
