+++ spark/dl/src/test/scala/com/intel/analytics/bigdl/nn/tf/AssignSpec.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.intel.analytics.bigdl.nn.ops
+package com.intel.analytics.bigdl.nn.tf
 
 import com.intel.analytics.bigdl.tensor.Tensor
 import com.intel.analytics.bigdl.utils.T
@@ -31,7 +31,7 @@ class AssignSpec extends FlatSpec with Matchers {
 
     val expectOutput = Tensor(T(2f, 2f, 4f))
 
-    val output = Assign().forward(input)
+    val output = new Assign().forward(input)
     output should be(expectOutput)
   }
 
@@ -45,7 +45,7 @@ class AssignSpec extends FlatSpec with Matchers {
 
     val expectOutput = Tensor(T(2.0f, 2.0f, 4.0f))
 
-    val output = Assign[Double]().forward(input)
+    val output = new Assign[Double]().forward(input)
     output should be(expectOutput)
   }
 }
